// elements/form-type.ts
export type FormData = {
    title?: string;
    firstName?: string;
    middleInitial?: string;
    lastName?: string;
    fullName?: string;
    company?: string;
    position?: string;
    addressLine1?: string;
    addressLine2?: string;
    city?: string;
    stateProvince?: string;
    country?: string;
    zip?: string;
    homePhone?: string;
    workPhone?: string;
    fax?: string;
    cellPhone?: string;
    email?: string;
    website?: string;
    userId?: string;
    password?: string;
    creditCardType?: string;
    creditCardNumber?: string;
    cardVerificationCode?: string;
    cardExpirationDate?: string;
    cardExpirationYear?: string;
    cardUserName?: string;
    cardIssuingBank?: string;
    cardCustomerServicePhone?: string;
    sex?: string;
    socialSecurityNumber?: string;
    driversLicenseNumber?: string;
    dayOfBirth?: string;
    monthOfBirth?: string;
    yearOfBirth?: string;
    age?: string;
    birthPlace?: string;
    income?: string;
    customMessage?: string;
    comments?: string;
  };